requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
beautifulsoup4==4.12.3
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.3
fastapi-cli==0.0.5
fastcore==1.7.19
fastlite==0.0.11
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.1.2
oauthlib==3.2.2
packaging==24.1
pandas==2.2.3
plotly==5.24.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-fasthtml==0.6.13
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
rich==13.9.3
scipy==1.14.1
shellingham==1.5.4
simpy==4.1.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlite-minutils==3.37.0.post3
starlette==0.41.0
tenacity==9.0.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.0
watchfiles==0.24.0
websockets==13.1
